make code better

+++ marketsvc/db/item.py
@@ -1,6 +1,5 @@
+from db.base import Base
 from sqlalchemy.orm import Mapped, mapped_column
-
-from db.base import Base, engine
 
 
 class Item(Base):
@@ -8,7 +7,7 @@ class Item(Base):
 
     id: Mapped[int] = mapped_column(primary_key=True)
     name: Mapped[str]
-    price: Mapped[int]
+    price: Mapped[float]
     description: Mapped[str | None]
 
     def __repr__(self) -> str:
